requirements.txt
alembic==1.6.4
APScheduler==3.6.3
certifi==2020.12.5
greenlet==1.1.0
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
tornado==6.1
tzlocal==2.1
